[BACKEND] ADD: Cleaning every routes and adding middleware that check if token is valdie
